Refactor: Simplified Xorn and Jolene implementation
