+++ Mage.Sets/src/mage/cards/x/Xorn.java
@@ -6,6 +6,7 @@ import mage.MageInt;
 import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
 import mage.abilities.effects.ReplacementEffectImpl;
+import mage.abilities.effects.common.ReplaceTreasureWithAdditionalEffect;
 import mage.constants.Duration;
 import mage.constants.Outcome;
 import mage.constants.SubType;
@@ -32,7 +33,7 @@ public final class Xorn extends CardImpl {
         this.toughness = new MageInt(2);
 
         // If you would create one or more Treasure tokens, instead create those tokens plus an additional Treasure token.
-        this.addAbility(new SimpleStaticAbility(new XornReplacementEffect()));
+        this.addAbility(new SimpleStaticAbility(new ReplaceTreasureWithAdditionalEffect()));
     }
 
     private Xorn(final Xorn card) {
@@ -42,59 +43,5 @@ public final class Xorn extends CardImpl {
     @Override
     public Xorn copy() {
         return new Xorn(this);
-    }
-}
-
-class XornReplacementEffect extends ReplacementEffectImpl {
-
-    public XornReplacementEffect() {
-        super(Duration.WhileOnBattlefield, Outcome.Benefit);
-        this.staticText = "If you would create one or more Treasure tokens, instead create those tokens plus an additional Treasure token";
-    }
-
-    private XornReplacementEffect(final XornReplacementEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public XornReplacementEffect copy() {
-        return new XornReplacementEffect(this);
-    }
-
-    @Override
-    public boolean checksEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.CREATE_TOKEN;
-    }
-
-    @Override
-    public boolean applies(GameEvent event, Ability source, Game game) {
-        if (event instanceof CreateTokenEvent && source.isControlledBy(event.getPlayerId())) {
-            for (Token token : ((CreateTokenEvent) event).getTokens().keySet()) {
-                if (token.hasSubtype(SubType.TREASURE, game)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public boolean replaceEvent(GameEvent event, Ability source, Game game) {
-        if (event instanceof CreateTokenEvent) {
-            CreateTokenEvent tokenEvent = (CreateTokenEvent) event;
-            TreasureToken treasureToken = null;
-            Map<Token, Integer> tokens = tokenEvent.getTokens();
-            for (Token token : tokens.keySet()) {
-                if (token instanceof TreasureToken) {
-                    treasureToken = (TreasureToken) token;
-                    break;
-                }
-            }
-            if (treasureToken == null) {
-                treasureToken = new TreasureToken();
-            }
-            tokens.put(treasureToken, tokens.getOrDefault(treasureToken, 0) + 1);
-        }
-        return false;
     }
 }
